Update spec interpretation
