+++ src/webgpu/api/operation/render_pipeline/pipeline_output_targets.spec.ts
@@ -3,9 +3,10 @@ export const description = `
 `;
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
-import { range, unreachable } from '../../../../common/util/util.js';
+import { range } from '../../../../common/util/util.js';
 import { kRenderableColorTextureFormats, kTextureFormatInfo } from '../../../capability_info.js';
 import { GPUTest } from '../../../gpu_test.js';
+import { getFragmentShaderCodeWithOutput } from '../../../util/shader.js';
 import { kTexelRepresentationInfo } from '../../../util/texture/texel_data.js';
 import { TexelView } from '../../../util/texture/texel_view.js';
 import { textureContentIsOKByT2B } from '../../../util/texture/texture_ok.js';
@@ -22,109 +23,7 @@ const kVertexShader = `
 }
 `;
 
-class F extends GPUTest {
-  // Build fragment shader based on output value and types
-  // e.g. write to color target 0 a vec4<f32>(1.0, 0.0, 1.0, 1.0) and color target 2 a vec2<u32>(1, 2)
-  // outputs: [
-  //   {
-  //     values: [1, 0, 1, 1],,
-  //     sampleType: 'float',
-  //     componentCount: 4,
-  //   },
-  //   null,
-  //   {
-  //     values: [1, 2],
-  //     sampleType: 'uint',
-  //     componentCount: 2,
-  //   },
-  // ]
-  //
-  // return:
-  // struct Outputs {
-  //     @location(0) o1 : vec4<f32>;
-  //     @location(2) o3 : vec2<u32>;
-  // }
-  // @stage(fragment) fn main() -> Outputs {
-  //     return Outputs(vec4<f32>(1.0, 0.0, 1.0, 1.0), vec4<u32>(1, 2));
-  // }
-  getFragmentShaderCode(
-    outputs: ({
-      values: readonly number[];
-      sampleType: GPUTextureSampleType;
-      componentCount: number;
-    } | null)[]
-  ): string {
-    const resultStrings = [] as string[];
-    let outputStructString = '';
-
-    for (let i = 0; i < outputs.length; i++) {
-      const o = outputs[i];
-      if (o === null) {
-        continue;
-      }
-
-      let fragColorType;
-      let suffix;
-      let fractionDigits = 0;
-      switch (o.sampleType) {
-        case 'sint':
-          fragColorType = 'i32';
-          suffix = '';
-          break;
-        case 'uint':
-          fragColorType = 'u32';
-          suffix = 'u';
-          break;
-        case 'float':
-        case 'unfilterable-float':
-          fragColorType = 'f32';
-          suffix = '';
-          fractionDigits = 4;
-          break;
-        default:
-          unreachable();
-      }
-
-      let outputType;
-      const v = o.values.map(n => n.toFixed(fractionDigits));
-      switch (o.componentCount) {
-        case 1:
-          outputType = fragColorType;
-          resultStrings.push(`${v[0]}${suffix}`);
-          break;
-        case 2:
-          outputType = `vec2<${fragColorType}>`;
-          resultStrings.push(`${outputType}(${v[0]}${suffix}, ${v[1]}${suffix})`);
-          break;
-        case 3:
-          outputType = `vec3<${fragColorType}>`;
-          resultStrings.push(`${outputType}(${v[0]}${suffix}, ${v[1]}${suffix}, ${v[2]}${suffix})`);
-          break;
-        case 4:
-          outputType = `vec4<${fragColorType}>`;
-          resultStrings.push(
-            `${outputType}(${v[0]}${suffix}, ${v[1]}${suffix}, ${v[2]}${suffix}, ${v[3]}${suffix})`
-          );
-          break;
-        default:
-          unreachable();
-      }
-
-      outputStructString += `@location(${i}) o${i} : ${outputType},\n`;
-    }
-
-    return `
-    struct Outputs {
-      ${outputStructString}
-    }
-
-    @stage(fragment) fn main() -> Outputs {
-        return Outputs(${resultStrings.join(',')});
-    }`;
-  }
-}
-
-export const g = makeTestGroup(F);
+export const g = makeTestGroup(GPUTest);
 
 // Values to write into each attachment
 // We make values different for each attachment index and each channel
@@ -181,7 +80,7 @@ g.test('color,attachments')
       },
       fragment: {
         module: t.device.createShaderModule({
-          code: t.getFragmentShaderCode(
+          code: getFragmentShaderCodeWithOutput(
             range(attachmentCount, i =>
               i === emptyAttachmentId
                 ? null
@@ -281,7 +180,7 @@ g.test('color,component_count')
       },
       fragment: {
         module: t.device.createShaderModule({
-          code: t.getFragmentShaderCode([
+          code: getFragmentShaderCodeWithOutput([
             {
               values,
               sampleType: info.sampleType,
@@ -491,7 +390,7 @@ The attachment has a load value of [1, 0, 0, 1]
       },
       fragment: {
         module: t.device.createShaderModule({
-          code: t.getFragmentShaderCode([
+          code: getFragmentShaderCodeWithOutput([
             {
               values: output,
               sampleType: info.sampleType,
