+++ src/webgpu/api/validation/createRenderPipeline.spec.ts
@@ -24,7 +24,7 @@ TODO: review existing tests, write descriptions, and make sure tests are complet
 `;
 
 import { makeTestGroup } from '../../../common/framework/test_group.js';
-import { unreachable } from '../../../common/util/util.js';
+import { assert, unreachable } from '../../../common/util/util.js';
 import {
   kTextureFormats,
   kRenderableColorTextureFormats,
@@ -35,60 +35,13 @@ import {
   kBlendFactors,
   kBlendOperations,
 } from '../../capability_info.js';
+import { getFragmentShaderCodeWithOutput } from '../../util/shader.js';
 import { kTexelRepresentationInfo } from '../../util/texture/texel_data.js';
 
 import { ValidationTest } from './validation_test.js';
 
+const values = [0, 1, 0, 1];
 class F extends ValidationTest {
-  getFragmentShaderCode(sampleType: GPUTextureSampleType, componentCount: number): string {
-    const v = ['0', '1', '0', '1'];
-
-    let fragColorType;
-    let suffix;
-    switch (sampleType) {
-      case 'sint':
-        fragColorType = 'i32';
-        suffix = '';
-        break;
-      case 'uint':
-        fragColorType = 'u32';
-        suffix = 'u';
-        break;
-      default:
-        fragColorType = 'f32';
-        suffix = '.0';
-        break;
-    }
-
-    let outputType;
-    let result;
-    switch (componentCount) {
-      case 1:
-        outputType = fragColorType;
-        result = `${v[0]}${suffix}`;
-        break;
-      case 2:
-        outputType = `vec2<${fragColorType}>`;
-        result = `${outputType}(${v[0]}${suffix}, ${v[1]}${suffix})`;
-        break;
-      case 3:
-        outputType = `vec3<${fragColorType}>`;
-        result = `${outputType}(${v[0]}${suffix}, ${v[1]}${suffix}, ${v[2]}${suffix})`;
-        break;
-      case 4:
-        outputType = `vec4<${fragColorType}>`;
-        result = `${outputType}(${v[0]}${suffix}, ${v[1]}${suffix}, ${v[2]}${suffix}, ${v[3]}${suffix})`;
-        break;
-      default:
-        unreachable();
-    }
-
-    return `
-    @stage(fragment) fn main() -> @location(0) ${outputType} {
-      return ${result};
-    }`;
-  }
-
   getDescriptor(
     options: {
       topology?: GPUPrimitiveTopology;
@@ -105,10 +58,13 @@ class F extends ValidationTest {
       targets = defaultTargets,
       sampleCount = 1,
       depthStencil,
-      fragmentShaderCode = this.getFragmentShaderCode(
-        kTextureFormatInfo[targets[0] ? targets[0].format : 'rgba8unorm'].sampleType,
-        4
-      ),
+      fragmentShaderCode = getFragmentShaderCodeWithOutput([
+        {
+          values,
+          sampleType: kTextureFormatInfo[targets[0] ? targets[0].format : 'rgba8unorm'].sampleType,
+          componentCount: 4,
+        },
+      ]),
       noFragment = false,
     } = options;
 
@@ -439,6 +395,11 @@ g.test('sample_count_must_be_valid')
     t.doCreateRenderPipelineTest(isAsync, _success, descriptor);
   });
 
+const kRenderableColorTextureFormatsExtend = [
+  undefined,
+  ...kRenderableColorTextureFormats,
+] as const;
+
 g.test('pipeline_output_targets')
   .desc(
     `Pipeline fragment output types must be compatible with target color state format
@@ -455,34 +416,51 @@ g.test('pipeline_output_targets')
     u
       .combine('isAsync', [false, true])
       .combine('writeMask', [0, 0x1, 0x2, 0x4, 0x8, 0xf])
-      .combine('format', kRenderableColorTextureFormats)
+      .combine('format', kRenderableColorTextureFormatsExtend)
       .beginSubcases()
+      .combine('hasShaderOutput', [false, true])
+      .filter(p => p.format === undefined || p.hasShaderOutput === true)
       .combine('sampleType', ['float', 'uint', 'sint'] as const)
       .combine('componentCount', [1, 2, 3, 4])
   )
   .beforeAllSubcases(t => {
     const { format } = t.params;
-    const info = kTextureFormatInfo[format];
-    t.selectDeviceOrSkipTestCase(info.feature);
+    if (format) {
+      const info = kTextureFormatInfo[format];
+      t.selectDeviceOrSkipTestCase(info.feature);
+    }
   })
   .fn(async t => {
-    const { isAsync, writeMask, format, sampleType, componentCount } = t.params;
-    const info = kTextureFormatInfo[format];
+    const { isAsync, writeMask, format, hasShaderOutput, sampleType, componentCount } = t.params;
+    const info = format ? kTextureFormatInfo[format] : null;
 
     const descriptor = t.getDescriptor({
-      targets: [{ format, writeMask }],
-      fragmentShaderCode: t.getFragmentShaderCode(sampleType, componentCount),
+      targets: format ? [{ format, writeMask }] : [],
+      // To have a dummy depthStencil attachment to avoid having no attachment at all which is invalid
+      depthStencil: { format: 'depth24plus' },
+      fragmentShaderCode: getFragmentShaderCodeWithOutput(
+        hasShaderOutput ? [{ values, sampleType, componentCount }] : []
+      ),
     });
 
-    const sampleTypeSuccess =
-      info.sampleType === 'float' || info.sampleType === 'unfilterable-float'
-        ? sampleType === 'float'
-        : info.sampleType === sampleType;
+    let _success = true;
+    if (hasShaderOutput) {
+      if (info) {
+        // there is a target correspond to the pipeline output
+        assert(format !== undefined);
+        const sampleTypeSuccess =
+          info.sampleType === 'float' || info.sampleType === 'unfilterable-float'
+            ? sampleType === 'float'
+            : info.sampleType === sampleType;
+        _success =
+          sampleTypeSuccess &&
+          componentCount >= kTexelRepresentationInfo[format].componentOrder.length;
+      } else {
+        // there is no target correspond to the pipeline output
+        _success = writeMask === 0;
+      }
+    }
 
-    const _success =
-      writeMask === 0 ||
-      (sampleTypeSuccess &&
-        componentCount >= kTexelRepresentationInfo[format].componentOrder.length);
     t.doCreateRenderPipelineTest(isAsync, _success, descriptor);
   });
 
@@ -587,7 +565,7 @@ g.test('pipeline_output_targets,blend')
           },
         },
       ],
-      fragmentShaderCode: t.getFragmentShaderCode(sampleType, componentCount),
+      fragmentShaderCode: getFragmentShaderCodeWithOutput([{ values, sampleType, componentCount }]),
     });
 
     const colorBlendReadsSrcAlpha =
@@ -638,7 +616,9 @@ Tests if blending is used, the target's format must be blendable (support "float
             },
           },
         ],
-        fragmentShaderCode: t.getFragmentShaderCode('float', 4),
+        fragmentShaderCode: getFragmentShaderCodeWithOutput([
+          { values, sampleType: 'float', componentCount: 4 },
+        ]),
       })
     );
   });
@@ -673,7 +653,9 @@ g.test('pipeline_output_targets,blend_min_max')
       dstFactor,
       operation,
     };
-    const fragmentShaderCode = t.getFragmentShaderCode('float', 4);
+    const fragmentShaderCode = getFragmentShaderCodeWithOutput([
+      { values, sampleType: 'float', componentCount: 4 },
+    ]);
     const format = 'rgba8unorm';
 
     const descriptor = t.getDescriptor({
@@ -725,7 +707,11 @@ g.test('pipeline_layout,device_mismatch')
         entryPoint: 'main',
       },
       fragment: {
-        module: t.device.createShaderModule({ code: t.getFragmentShaderCode('float', 4) }),
+        module: t.device.createShaderModule({
+          code: getFragmentShaderCodeWithOutput([
+            { values, sampleType: 'float', componentCount: 4 },
+          ]),
+        }),
         entryPoint: 'main',
         targets: [{ format: 'rgba8unorm' }] as const,
       },
@@ -766,8 +752,16 @@ g.test('shader_module,device_mismatch')
       },
       fragment: {
         module: fragment_mismatched
-          ? t.mismatchedDevice.createShaderModule({ code: t.getFragmentShaderCode('float', 4) })
-          : t.device.createShaderModule({ code: t.getFragmentShaderCode('float', 4) }),
+          ? t.mismatchedDevice.createShaderModule({
+              code: getFragmentShaderCodeWithOutput([
+                { values, sampleType: 'float', componentCount: 4 },
+              ]),
+            })
+          : t.device.createShaderModule({
+              code: getFragmentShaderCodeWithOutput([
+                { values, sampleType: 'float', componentCount: 4 },
+              ]),
+            }),
         entryPoint: 'main',
         targets: [{ format: 'rgba8unorm' }] as const,
       },
